Model-API2/requirements.txt
flask
tensorflow
pillow
numpy
google-cloud-storage
